+++ test/units/Transport/Configuration/ConfigurationTest.php
@@ -22,11 +22,7 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
     {
         $conf = new Configuration();
 
-        $this->assertArrayHasKey('allow_redirects', $conf);
-        $this->assertArrayHasKey('user_agent', $conf);
-        $this->assertArrayHasKey('allow_redirects', $conf);
         $this->assertArrayHasKey('connect_timeout', $conf);
-        $this->assertArrayHasKey('http_errors', $conf);
         $this->assertArrayHasKey('timeout', $conf);
         $this->assertArrayHasKey('verify', $conf);
         $this->assertArrayHasKey('url', $conf);
@@ -48,15 +44,6 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(120, $conf->timeout);
     }
 
-    public function testHttpErrors()
-    {
-        $conf = new Configuration();
-
-        $this->assertFalse($conf->http_errors);
-        $conf->http_errors = true;
-        $this->assertTrue($conf->http_errors);
-    }
-
     public function testSslVerify()
     {
         $conf = new Configuration();
@@ -64,19 +51,5 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($conf->verify);
         $conf->verify = true;
         $this->assertTrue($conf->verify);
-    }
-
-    public function testAllowRedirect()
-    {
-        $conf = new Configuration(['allow_redirects' => false]);
-        $this->assertFalse($conf->redirectsAllowed());
-
-        $this->assertNull($conf->allowRedirectsMax());
-        $conf->allowRedirectsMax(1);
-        $this->assertTrue($conf->redirectsAllowed());
-        $this->assertEquals(1,$conf->allowRedirectsMax());
-        $this->assertTrue($conf->allowRedirectsReferer());
-        $conf->allowRedirectsReferer(false);
-        $this->assertFalse($conf->allowRedirectsReferer());
     }
 }
